requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
bidict==0.22.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
colorzero==1.1
contourpy==1.1.1
cryptography==3.3.2
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
flask==3.0.0
Flask-SocketIO==5.3.6
gpiozero==1.6.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.25.1
RPi.GPIO==0.7.0
simple-websocket==1.0.0
smbus==1.1.post2
toml==0.10.2
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
webencodings==0.5.1
werkzeug==3.0.0
wrapt==1.12.1
wsproto==1.2.0